requirements.txt
streamlit
tensorflow
pillow
google-generativeai
fpdf2
numpy
